Refactor: Create Referee class
- to handle checkState, checkVertical, checkHorizontal, and checkDiagonal functions

- for improve code readability and test
