+++ app/src/main/java/nextstep/omok/MainActivity.kt
@@ -1,5 +1,6 @@
 package nextstep.omok
 
+import Referee
 import android.os.Bundle
 import android.widget.ImageView
 import android.widget.TableLayout
@@ -15,10 +16,13 @@ class MainActivity : AppCompatActivity() {
 
     private val boardSize = 15
     private var boardState = Array<Int?>(boardSize * boardSize) { null }
+    private lateinit var referee: Referee
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+
+        referee = Referee(boardSize, boardState)
 
         val board = findViewById<TableLayout>(R.id.board)
         initializeBoard(board)
@@ -54,49 +58,9 @@ class MainActivity : AppCompatActivity() {
         boardState[position] = 1   //white
         currentStone = 0}
 
-        if (checkState(position)) {displayResult()}
-    }
-
-    private fun checkState(position: Int): Boolean{
-        val row = position / boardSize
-        val column = position % boardSize
-        val stone = boardState[position]?: return false
-
-        return checkVertical(row,column,stone) || checkHorizontal(row,column,stone) || checkDiagonal(row,column,stone)
-    }
-
-    private fun checkVertical(row: Int, column: Int, stone: Int): Boolean{
-        val topCount = (row - 1 downTo 0).takeWhile { boardState[it * boardSize + column] == stone }.count()
-        val bottomCount = (row + 1 until boardSize).takeWhile { boardState[it * boardSize + column] == stone }.count()
-
-        val count = topCount + bottomCount
-
-        return (count >= 4)
-    }
-
-    private fun checkHorizontal(row: Int, column: Int, stone: Int): Boolean{
-
-        val leftCount = (column - 1 downTo 0).takeWhile { boardState[row * boardSize + it] == stone }.count()
-        val rightCount = (column + 1 until boardSize).takeWhile { boardState[row * boardSize + it] == stone }.count()
-
-        val count = leftCount + rightCount
-
-        return (count >= 4)
-    }
-
-    private fun checkDiagonal(row: Int, column: Int, stone: Int): Boolean{
-        // 왼->오 대각선
-        val leftTopCount = (1..minOf(row, column)).takeWhile { boardState[(row - it) * boardSize + (column - it)] == stone }.count()
-        val rightBottomCount = (1..minOf(boardSize - row - 1, boardSize - column - 1)).takeWhile { boardState[(row + it) * boardSize + (column + it)] == stone }.count()
-
-        // 오->왼 대각선
-        val rightTopCount = (1..minOf(row, boardSize - column - 1)).takeWhile { boardState[(row - it) * boardSize + (column + it)] == stone }.count()
-        val leftBottomCount = (1..minOf(boardSize - row - 1, column)).takeWhile { boardState[(row + it) * boardSize + (column - it)] == stone }.count()
-
-        val countd1 = leftTopCount + rightBottomCount
-        val countd2 = rightTopCount + leftBottomCount
-
-        return (countd1>=4 || countd2>=4)
+        if (referee.checkState(position)) {
+            displayResult()
+        }
     }
 
     private fun displayResult() {
